LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpReceiver-getParameters_rest-expected.txt


FAIL getParameters() with simulcast video receiver assert_greater_than: expected a number greater than 0 but got 0
